Extract from_google + to_google
