+++ caldir-provider-google/src/google/mod.rs
@@ -1,2 +1,4 @@
+pub mod actions;
 pub mod api;
-pub mod index;
+pub mod from_google;
+pub mod to_google;
